Database notifications to string
